requirements.txt
# MkDocs Core
mkdocs==1.5.3

# Тема Material
mkdocs-material==9.5.15

# Плагины
mkdocs-glightbox==0.4.0
mkdocs-mermaid2-plugin==1.1.1
mkdocs-git-revision-date-localized-plugin==1.2.2
mkdocs-minify-plugin==0.7.2

# Markdown расширения
pymdown-extensions==10.7
pygments==2.17.2

# Для синхронизации
GitPython>=3.1.40
PyYAML>=6.0.1
requests>=2.31.0